requirements.txt
python-telegram-bot[job-queue]==21.7
gspread==6.1.4
google-auth==2.36.0
boto3==1.35.86
pydantic-settings==2.7.0
python-dotenv==1.0.1
